t/unit-tests: implement clar specific reftable helper functions
Helper functions defined in `t/unit-tests/lib-reftable.{c,h}` are
required for the reftable-related test files to run efficeintly. In the
current implementation these functions are designed to conform with our
homegrown unit-testing structure. So in other to convert the reftable
test files, there is need for a clar specific implementation of these
helper functions.

Adapt functions in lib-reftable.{c,h} to use clar. These functions are
adapted to conform with the clar testing framework and become available
for all reftable-related test files implemented using the clar testing
framework, which requires them. This will be used by subsequent commits.